# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Penyimpanan Gambar",
        "description": "Dokumentasi API untuk layanan penyimpanan dan pengambilan gambar.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/images": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Mengambil daftar gambar",
                "description": "Mengembalikan daftar semua gambar yang ada, dengan opsi pencarian berdasarkan judul.",
                "operationId": "getImagesList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci untuk mencari gambar berdasarkan judul.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Image"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Images"
                ],
                "summary": "Mengunggah gambar baru",
                "description": "Menyimpan file gambar baru beserta judulnya ke dalam sistem.",
                "operationId": "storeImage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Judul untuk gambar.",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "File gambar yang akan diunggah.",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Gambar berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Image"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/images/{id}": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Mengambil data satu gambar",
                "description": "Mengembalikan detail data gambar berdasarkan ID.",
                "operationId": "getImageById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Gambar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Image"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Images"
                ],
                "summary": "Menghapus gambar",
                "description": "Menghapus file gambar dari storage dan data dari database berdasarkan ID.",
                "operationId": "deleteImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Gambar yang akan dihapus",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gambar berhasil dihapus",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Image deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Image": {
                "title": "Image",
                "description": "Model Gambar",
                "properties": {
                    "id": {
                        "description": "ID unik gambar",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Judul gambar",
                        "type": "string"
                    },
                    "file_path": {
                        "description": "Path lokasi file gambar",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Waktu dibuat",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Waktu diperbarui",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API Key untuk otentikasi.",
                "name": "X-API-Key",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Images",
            "description": "Images"
        }
    ]
}